git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C705

+++ src/Assets/Data/Effects/os_default.fx
@@ -90,7 +90,7 @@ float4 PS_Main(PS_INPUT input) : SV_TARGET
 
 	//lighting *= (LightRange / dot(input.LightDir, input.LightDir));
 	
-	for (int i = 0; i < LightsCount; i++)
+	for (float i = 0; i < LightsCount; i++)
 	{
 		float lighting = dot(
 			input.Normal,
